Move transformation_manager.hpp to include folder (to be used in rviz_default_plugins)
