+++ rviz_common/CMakeLists.txt
@@ -133,7 +133,7 @@ set(rviz_common_headers_to_moc
   include/rviz_common/tool.hpp
   src/rviz_common/tool_manager.hpp
   src/rviz_common/tool_properties_panel.hpp
-  src/rviz_common/transformation/transformation_manager.hpp
+  include/rviz_common/transformation/transformation_manager.hpp
   src/rviz_common/transformation_panel.hpp
   include/rviz_common/view_controller.hpp
   include/rviz_common/view_manager.hpp
