+++ rviz_common/src/rviz_common/visualization_manager.hpp
@@ -44,7 +44,7 @@
 #include "rviz_common/display_context.hpp"
 #include "rviz_common/frame_manager_iface.hpp"
 #include "rviz_common/ros_integration/ros_node_abstraction_iface.hpp"
-#include "transformation/transformation_manager.hpp"
+#include "rviz_common/transformation/transformation_manager.hpp"
 
 class QTimer;
